Coins
So, I added to the game the possibility of collecting coins. Animation of coins, collection and accrual of points and if the player collects one of the coins then he removes himself from the map and there is no possibility to collect it a second time. I still need to learn how to use signals in godot is a very interesting topic for communication between nodes.