Main Menu and End Game
I added a main menu to the game where there is a choice of either we can play or exit the game, I also completed the second level of the game and after entering the flag on the second level the game shuts down. Perhaps I will add an end screen where there will be an option to exit the game.